move the config submit button up to the SettingsTab component
